补全缺失的 Blueprint 功能并修复编译问题
- 添加 HTTP Blueprint 功能 (使用 Latent Actions 实现异步请求)
  - HttpGet: 异步 GET 请求
  - HttpPost: 异步 POST 请求
  - HttpDownloadFile: 异步文件下载
- 添加 FileIO 缺失的 Blueprint 函数
  - WriteFileLines, ReadBinaryFile, WriteBinaryFile
  - DeleteDirectoryAtPath, MoveFileToPath
  - GetDirectoriesInDir, GetFileDir
- 添加 JSON Map 相关 Blueprint 函数
  - MapToJson, JsonToMap
- 修复 SHA256 实现: 使用 Windows CryptoAPI 实现真正的 SHA256
- 修复 ClipboardHelper 32位移位警告

Author: lostpanda

+++ MySuperTools/Plugins/SuperTools/Source/SuperTools/Private/Utils/ClipboardHelper.cpp
@@ -214,7 +214,7 @@ bool FClipboardHelper::GetImageDataFromClipboard(TArray<uint8>& OutImageData, in
 	int32 ColorTableSize = 0;
 	if (pBitmapHeader->biBitCount <= 8)
 	{
-		ColorTableSize = (1 << pBitmapHeader->biBitCount) * sizeof(RGBQUAD);
+		ColorTableSize = static_cast<int32>((1ULL << pBitmapHeader->biBitCount) * sizeof(RGBQUAD));
 	}
 
 	uint8* pPixelData = (uint8*)pBitmapHeader + pBitmapHeader->biSize + ColorTableSize;
